Moving the backend unit tests to the respective backend directory
