--- tests/unit/test_sglang_unit.py
@@ -1,75 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-"""Unit tests for SGLang backend components."""
-
-import re
-import sys
-from pathlib import Path
-
-import pytest
-
-from dynamo.sglang.args import parse_args
-from tests.unit.conftest import make_cli_args_fixture
-
-# Get path relative to this test file
-TEST_DIR = Path(__file__).parent.parent
-JINJA_TEMPLATE_PATH = str(TEST_DIR / "serve" / "fixtures" / "custom_template.jinja")
-
-
-pytestmark = [
-    pytest.mark.unit,
-    pytest.mark.sglang,
-    pytest.mark.gpu_1,
-]
-
-
-# Create SGLang-specific CLI args fixture
-# This will use monkeypatch to write to argv
-mock_sglang_cli = make_cli_args_fixture("dynamo.sglang")
-
-
-@pytest.mark.asyncio
-async def test_custom_jinja_template_invalid_path(mock_sglang_cli):
-    """Test that invalid file path raises FileNotFoundError."""
-    invalid_path = "/nonexistent/path/to/template.jinja"
-    mock_sglang_cli(
-        "--model", "Qwen/Qwen3-0.6B", "--custom-jinja-template", invalid_path
-    )
-
-    with pytest.raises(
-        FileNotFoundError,
-        match=re.escape(f"Custom Jinja template file not found: {invalid_path}"),
-    ):
-        await parse_args(sys.argv[1:])
-
-
-@pytest.mark.asyncio
-async def test_custom_jinja_template_valid_path(mock_sglang_cli):
-    """Test that valid absolute path is stored correctly."""
-    mock_sglang_cli(model="Qwen/Qwen3-0.6B", custom_jinja_template=JINJA_TEMPLATE_PATH)
-
-    config = await parse_args(sys.argv[1:])
-
-    assert config.dynamo_args.custom_jinja_template == JINJA_TEMPLATE_PATH, (
-        f"Expected custom_jinja_template value to be {JINJA_TEMPLATE_PATH}, "
-        f"got {config.dynamo_args.custom_jinja_template}"
-    )
-
-
-@pytest.mark.asyncio
-async def test_custom_jinja_template_env_var_expansion(monkeypatch, mock_sglang_cli):
-    """Test that environment variables in paths are expanded by Python code."""
-    jinja_dir = str(TEST_DIR / "serve" / "fixtures")
-    monkeypatch.setenv("JINJA_DIR", jinja_dir)
-
-    cli_path = "$JINJA_DIR/custom_template.jinja"
-    mock_sglang_cli(model="Qwen/Qwen3-0.6B", custom_jinja_template=cli_path)
-
-    config = await parse_args(sys.argv[1:])
-
-    assert "$JINJA_DIR" not in config.dynamo_args.custom_jinja_template
-    assert config.dynamo_args.custom_jinja_template == JINJA_TEMPLATE_PATH, (
-        f"Expected custom_jinja_template value to be {JINJA_TEMPLATE_PATH}, "
-        f"got {config.dynamo_args.custom_jinja_template}"
-    )
